requirements.txt
tensorflow
numpy
opencv-python
matplotlib
pandas
scikit-learn
imbalanced-learn
streamlit
pillow
lxml
